rename services.js to userServices.js
